Functional Tests : API (Refactoring directories / Rename file / Add blocked UI Tests)
