--- tests/UI/campaigns/functional/API/02_endpoints/01_apiClient/02_deleteAPIApiClientId.ts
@@ -1,170 +1,0 @@
-// Import utils
-import api from '@utils/api';
-import helper from '@utils/helpers';
-import testContext from '@utils/testContext';
-
-// Import commonTests
-import loginCommon from '@commonTests/BO/loginBO';
-
-// Import pages
-import apiClientPage from 'pages/BO/advancedParameters/APIClient';
-import addNewApiClientPage from '@pages/BO/advancedParameters/APIClient/add';
-import dashboardPage from '@pages/BO/dashboard';
-
-// Import data
-import APIClientData from '@data/faker/APIClient';
-
-import {expect} from 'chai';
-import type {APIRequestContext, BrowserContext, Page} from 'playwright';
-
-const baseContext: string = 'functional_API_endpoints_apiClient_deleteAPIApiClientId';
-
-describe('API : DELETE /api-client/{apiClientId}', async () => {
-  let apiContext: APIRequestContext;
-  let browserContext: BrowserContext;
-  let page: Page;
-  let accessToken: string;
-  let clientSecret: string;
-  let idApiClient: number;
-
-  const clientScope: string = 'api_client_write';
-  const clientData: APIClientData = new APIClientData({
-    enabled: true,
-    scopes: [
-      clientScope,
-    ],
-  });
-
-  before(async function () {
-    browserContext = await helper.createBrowserContext(this.browser);
-    page = await helper.newTab(browserContext);
-
-    apiContext = await helper.createAPIContext(global.API.URL);
-  });
-
-  after(async () => {
-    await helper.closeBrowserContext(browserContext);
-  });
-
-  describe('BackOffice : Fetch the access token', async () => {
-    it('should login in BO', async function () {
-      await loginCommon.loginBO(this, page);
-    });
-
-    it('should go to \'Advanced Parameters > API Client\' page', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'goToAdminAPIPage', baseContext);
-
-      await dashboardPage.goToSubMenu(
-        page,
-        dashboardPage.advancedParametersLink,
-        dashboardPage.adminAPILink,
-      );
-
-      const pageTitle = await apiClientPage.getPageTitle(page);
-      expect(pageTitle).to.eq(apiClientPage.pageTitle);
-    });
-
-    it('should check that no records found', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'checkThatNoRecordFound', baseContext);
-
-      const noRecordsFoundText = await apiClientPage.getTextForEmptyTable(page);
-      expect(noRecordsFoundText).to.contains('warning No records found');
-    });
-
-    it('should go to add New API Client page', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'goToNewAPIClientPage', baseContext);
-
-      await apiClientPage.goToNewAPIClientPage(page);
-
-      const pageTitle = await addNewApiClientPage.getPageTitle(page);
-      expect(pageTitle).to.eq(addNewApiClientPage.pageTitleCreate);
-    });
-
-    it('should create API Client', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'createAPIClient', baseContext);
-
-      const textResult = await addNewApiClientPage.addAPIClient(page, clientData);
-      expect(textResult).to.contains(addNewApiClientPage.successfulCreationMessage);
-
-      const textMessage = await addNewApiClientPage.getAlertInfoBlockParagraphContent(page);
-      expect(textMessage).to.contains(addNewApiClientPage.apiClientGeneratedMessage);
-    });
-
-    it('should copy client secret', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'copyClientSecret', baseContext);
-
-      await addNewApiClientPage.copyClientSecret(page);
-
-      clientSecret = await addNewApiClientPage.getClipboardText(page);
-      expect(clientSecret.length).to.be.gt(0);
-    });
-
-    it('should request the endpoint /access_token', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'requestOauth2Token', baseContext);
-
-      const apiResponse = await apiContext.post('access_token', {
-        form: {
-          client_id: clientData.clientId,
-          client_secret: clientSecret,
-          grant_type: 'client_credentials',
-          scope: clientScope,
-        },
-      });
-      expect(apiResponse.status()).to.eq(200);
-      expect(api.hasResponseHeader(apiResponse, 'Content-Type')).to.eq(true);
-      expect(api.getResponseHeader(apiResponse, 'Content-Type')).to.contains('application/json');
-
-      const jsonResponse = await apiResponse.json();
-      expect(jsonResponse).to.have.property('access_token');
-      expect(jsonResponse.token_type).to.be.a('string');
-
-      accessToken = jsonResponse.access_token;
-    });
-  });
-
-  describe('BackOffice : Fetch the ID of the API Client', async () => {
-    it('should go to \'Advanced Parameters > API Client\' page', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'returnToAdminAPIPage', baseContext);
-
-      await dashboardPage.goToSubMenu(
-        page,
-        dashboardPage.advancedParametersLink,
-        dashboardPage.adminAPILink,
-      );
-
-      const pageTitle = await apiClientPage.getPageTitle(page);
-      expect(pageTitle).to.eq(apiClientPage.pageTitle);
-    });
-
-    it('should get informations', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'getInformations', baseContext);
-
-      idApiClient = parseInt(await apiClientPage.getTextColumn(page, 'id_api_client', 1), 10);
-      expect(idApiClient).to.be.gt(0);
-    });
-  });
-
-  describe('API : Delete the API Access', async () => {
-    it('should request the endpoint /api-client/{apiClientId}', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'requestEndpoint', baseContext);
-
-      const apiResponse = await apiContext.delete(`api-client/${idApiClient}`, {
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-        },
-      });
-      expect(apiResponse.status()).to.eq(204);
-    });
-  });
-
-  describe('BackOffice : Check the API Access is deleted', async () => {
-    it('should check that the API Access is deleted ', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'filterAfterDeletion', baseContext);
-
-      await apiClientPage.reloadPage(page);
-
-      const numberOfGroupsAfterCreation = await apiClientPage.getNumberOfElementInGrid(page);
-      expect(numberOfGroupsAfterCreation).to.be.equal(0);
-    });
-  });
-});
