+++ tests/UI/campaigns/functional/API/02_endpoints/01_apiClient/02_deleteApiClientId.ts
@@ -4,7 +4,6 @@ import helper from '@utils/helpers';
 import testContext from '@utils/testContext';
 
 // Import commonTests
-import {deleteAPIClientTest} from '@commonTests/BO/advancedParameters/authServer';
 import loginCommon from '@commonTests/BO/loginBO';
 
 // Import pages
@@ -18,18 +17,17 @@ import APIClientData from '@data/faker/APIClient';
 import {expect} from 'chai';
 import type {APIRequestContext, BrowserContext, Page} from 'playwright';
 
-const baseContext: string = 'functional_API_endpoints_apiClient_getAPIApiClientId';
+const baseContext: string = 'functional_API_endpoints_apiClient_deleteApiClientId';
 
-describe('API : GET /api-client/{apiClientId}', async () => {
+describe('API : DELETE /api-client/{apiClientId}', async () => {
   let apiContext: APIRequestContext;
   let browserContext: BrowserContext;
   let page: Page;
   let accessToken: string;
-  let jsonResponse: any;
   let clientSecret: string;
   let idApiClient: number;
 
-  const clientScope: string = 'api_client_read';
+  const clientScope: string = 'api_client_write';
   const clientData: APIClientData = new APIClientData({
     enabled: true,
     scopes: [
@@ -124,7 +122,7 @@ describe('API : GET /api-client/{apiClientId}', async () => {
     });
   });
 
-  describe('BackOffice : Expected data', async () => {
+  describe('BackOffice : Fetch the ID of the API Client', async () => {
     it('should go to \'Advanced Parameters > API Client\' page', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'returnToAdminAPIPage', baseContext);
 
@@ -146,101 +144,27 @@ describe('API : GET /api-client/{apiClientId}', async () => {
     });
   });
 
-  describe('API : Check Data', async () => {
+  describe('API : Delete the API Access', async () => {
     it('should request the endpoint /api-client/{apiClientId}', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'requestEndpoint', baseContext);
 
-      const apiResponse = await apiContext.get(`api-client/${idApiClient}`, {
+      const apiResponse = await apiContext.delete(`api-client/${idApiClient}`, {
         headers: {
           Authorization: `Bearer ${accessToken}`,
         },
       });
-      expect(apiResponse.status()).to.eq(200);
-      expect(api.hasResponseHeader(apiResponse, 'Content-Type')).to.eq(true);
-      expect(api.getResponseHeader(apiResponse, 'Content-Type')).to.contains('application/json');
-
-      jsonResponse = await apiResponse.json();
-    });
-
-    it('should check the JSON Response keys', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'checkResponseKeys', baseContext);
-
-      expect(jsonResponse).to.have.all.keys(
-        'apiClientId',
-        'clientId',
-        'clientName',
-        'description',
-        'externalIssuer',
-        'enabled',
-        'lifetime',
-        'scopes',
-      );
-    });
-
-    it('should check the JSON Response : `apiClientId`', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'checkResponseApiClientId', baseContext);
-
-      expect(jsonResponse).to.have.property('apiClientId');
-      expect(jsonResponse.apiClientId).to.be.a('number');
-      expect(jsonResponse.apiClientId).to.be.equal(idApiClient);
-    });
-
-    it('should check the JSON Response : `clientId`', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'checkResponseClientId', baseContext);
-
-      expect(jsonResponse).to.have.property('clientId');
-      expect(jsonResponse.clientId).to.be.a('string');
-      expect(jsonResponse.clientId).to.be.equal(clientData.clientId);
-    });
-
-    it('should check the JSON Response : `clientName`', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'checkResponseClientName', baseContext);
-
-      expect(jsonResponse).to.have.property('clientName');
-      expect(jsonResponse.clientName).to.be.a('string');
-      expect(jsonResponse.clientName).to.be.equal(clientData.clientName);
-    });
-
-    it('should check the JSON Response : `description`', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'checkResponseDescription', baseContext);
-
-      expect(jsonResponse).to.have.property('description');
-      expect(jsonResponse.description).to.be.a('string');
-      expect(jsonResponse.description).to.be.equal(clientData.description);
-    });
-
-    it('should check the JSON Response : `externalIssuer`', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'checkResponseExternalIssuer', baseContext);
-
-      expect(jsonResponse).to.have.property('externalIssuer');
-      expect(jsonResponse.externalIssuer).to.equal(null);
-    });
-
-    it('should check the JSON Response : `enabled`', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'checkResponseEnabled', baseContext);
-
-      expect(jsonResponse).to.have.property('enabled');
-      expect(jsonResponse.enabled).to.be.a('boolean');
-      expect(jsonResponse.enabled).to.be.equal(clientData.enabled);
-    });
-
-    it('should check the JSON Response : `lifetime`', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'checkResponseLifetime', baseContext);
-
-      expect(jsonResponse).to.have.property('lifetime');
-      expect(jsonResponse.lifetime).to.be.a('number');
-      expect(jsonResponse.lifetime).to.be.equal(clientData.tokenLifetime);
-    });
-
-    it('should check the JSON Response : `scopes`', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'checkResponseScopes', baseContext);
-
-      expect(jsonResponse).to.have.property('scopes');
-      expect(jsonResponse.scopes).to.be.a('array');
-      expect(jsonResponse.scopes).to.deep.equal(clientData.scopes);
+      expect(apiResponse.status()).to.eq(204);
     });
   });
 
-  // Post-condition: Create an API Client
-  deleteAPIClientTest(`${baseContext}_postTest`);
+  describe('BackOffice : Check the API Access is deleted', async () => {
+    it('should check that the API Access is deleted ', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'filterAfterDeletion', baseContext);
+
+      await apiClientPage.reloadPage(page);
+
+      const numberOfGroupsAfterCreation = await apiClientPage.getNumberOfElementInGrid(page);
+      expect(numberOfGroupsAfterCreation).to.be.equal(0);
+    });
+  });
 });
