+++ tests/UI/campaigns/functional/API/02_endpoints/08_module/01_getModuleId.ts
@@ -15,25 +15,25 @@ import {moduleManager} from '@pages/BO/modules/moduleManager';
 
 // Import data
 import APIClientData from '@data/faker/APIClient';
-import type {ModuleInfo} from '@data/types/module';
+import {ModuleInfo} from '@data/types/module';
 
-import {
-  boModuleManagerPage,
-  FakerModule,
-} from '@prestashop-core/ui-testing';
 import {expect} from 'chai';
 import type {APIRequestContext, BrowserContext, Page} from 'playwright';
+import {
+  boModuleManagerPage,
+  dataModules,
+} from '@prestashop-core/ui-testing';
 
-const baseContext: string = 'functional_API_endpoints_module_getAPIModules';
+const baseContext: string = 'functional_API_endpoints_module_getModuleId';
 
-describe('API : GET /modules', async () => {
+describe('API : GET /module/{moduleId}', async () => {
   let apiContext: APIRequestContext;
   let browserContext: BrowserContext;
   let page: Page;
   let clientSecret: string;
   let accessToken: string;
   let jsonResponse: any;
-  const jsonResponseItems: ModuleInfo [] = [];
+  let moduleInfo: ModuleInfo;
 
   const clientScope: string = 'module_read';
   const clientData: APIClientData = new APIClientData({
@@ -130,98 +130,83 @@ describe('API : GET /modules', async () => {
     });
   });
 
-  describe('API : Fetch Data', async () => {
-    [
-      {
-        page: 0,
-      },
-      {
-        page: 1,
-      },
-    ].forEach((arg: {page: number}, index: number) => {
-      it(`should request the endpoint /modules (page ${arg.page})`, async function () {
-        await testContext.addContextItem(this, 'testIdentifier', `requestEndpoint${index}`, baseContext);
-
-        const apiResponse = await apiContext.get(`modules${arg.page > 0 ? `?offset=${arg.page * 50}` : ''}`, {
-          headers: {
-            Authorization: `Bearer ${accessToken}`,
-          },
-        });
-        expect(apiResponse.status()).to.eq(200);
-        expect(api.hasResponseHeader(apiResponse, 'Content-Type')).to.eq(true);
-        expect(api.getResponseHeader(apiResponse, 'Content-Type')).to.contains('application/json');
-
-        jsonResponse = await apiResponse.json();
-      });
-
-      it(`should check the JSON Response keys (page ${arg.page})`, async function () {
-        await testContext.addContextItem(this, 'testIdentifier', `checkResponseKeys${index}`, baseContext);
-
-        const keys = [
-          'totalItems',
-          'sortOrder',
-          'limit',
-          'filters',
-          'items',
-        ];
-
-        if (arg.page > 0) {
-          keys.push('offset');
-        }
-        expect(jsonResponse).to.have.all.keys(keys);
-
-        expect(jsonResponse.items.length).to.be.gt(0);
-        if (arg.page === 0) {
-          expect(jsonResponse.items.length).to.be.equal(jsonResponse.limit);
-        } else {
-          expect(jsonResponse.items.length).to.be.lt(jsonResponse.limit);
-        }
-
-        for (let i:number = 0; i < jsonResponse.items.length; i++) {
-          expect(jsonResponse.items[i]).to.have.all.keys(
-            'moduleId',
-            'technicalName',
-            'version',
-            'enabled',
-          );
-          jsonResponseItems.push(jsonResponse.items[i]);
-        }
-      });
-    });
-  });
-
   describe('BackOffice : Expected data', async () => {
     it('should go to \'Modules > Module Manager\' page', async function () {
       await testContext.addContextItem(this, 'testIdentifier', 'goToModulesPage', baseContext);
 
       await dashboardPage.goToSubMenu(page, dashboardPage.modulesParentLink, dashboardPage.modulesParentLink);
       await boModuleManagerPage.closeSfToolBar(page);
-      await moduleManager.filterByStatus(page, 'installed');
 
       const pageTitle = await boModuleManagerPage.getPageTitle(page);
       expect(pageTitle).to.contains(boModuleManagerPage.pageTitle);
-
-      const numModules = await moduleManager.getNumberOfModules(page);
-      expect(numModules).to.eq(jsonResponseItems.length);
     });
 
-    it('should filter list by technicaleName', async function () {
-      await testContext.addContextItem(this, 'testIdentifier', 'checkJSONItems', baseContext);
+    it('should filter list by name', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'searchModule', baseContext);
 
-      for (let idxItem: number = 0; idxItem < jsonResponseItems.length; idxItem++) {
-        // eslint-disable-next-line no-loop-func
-        const isModuleVisible = await moduleManager.searchModule(
-          page,
-          {tag: jsonResponseItems[idxItem].technicalName} as FakerModule,
-        );
-        expect(isModuleVisible).to.be.equal(true);
+      const isModuleVisible = await moduleManager.searchModule(page, dataModules.psApiResources);
+      expect(isModuleVisible).to.be.equal(true);
 
-        const moduleInfos = await moduleManager.getModuleInformationNth(page, 1);
-        expect(moduleInfos.moduleId).to.equal(jsonResponseItems[idxItem].moduleId);
-        expect(moduleInfos.technicalName).to.equal(jsonResponseItems[idxItem].technicalName);
-        expect(moduleInfos.version).to.equal(jsonResponseItems[idxItem].version);
-        expect(moduleInfos.enabled).to.equal(jsonResponseItems[idxItem].enabled);
-      }
+      moduleInfo = await moduleManager.getModuleInformationNth(page, 1);
+    });
+  });
+
+  describe('API : Check Data', async () => {
+    it('should request the endpoint /module/{moduleId}', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'requestEndpoint', baseContext);
+
+      const apiResponse = await apiContext.get(`module/${moduleInfo.moduleId}`, {
+        headers: {
+          Authorization: `Bearer ${accessToken}`,
+        },
+      });
+      expect(apiResponse.status()).to.eq(200);
+      expect(api.hasResponseHeader(apiResponse, 'Content-Type')).to.eq(true);
+      expect(api.getResponseHeader(apiResponse, 'Content-Type')).to.contains('application/json');
+
+      jsonResponse = await apiResponse.json();
+    });
+
+    it('should check the JSON Response keys', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'checkResponseKeys', baseContext);
+      expect(jsonResponse).to.have.all.keys(
+        'moduleId',
+        'technicalName',
+        'version',
+        'enabled',
+      );
+    });
+
+    it('should check the JSON Response : `moduleId`', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'checkResponseModuleId', baseContext);
+
+      expect(jsonResponse).to.have.property('moduleId');
+      expect(jsonResponse.moduleId).to.be.a('number');
+      expect(jsonResponse.moduleId).to.be.equal(moduleInfo.moduleId);
+    });
+
+    it('should check the JSON Response : `technicalName`', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'checkResponseTechnicalName', baseContext);
+
+      expect(jsonResponse).to.have.property('technicalName');
+      expect(jsonResponse.technicalName).to.be.a('string');
+      expect(jsonResponse.technicalName).to.be.equal(moduleInfo.technicalName);
+    });
+
+    it('should check the JSON Response : `version`', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'checkResponseVersion', baseContext);
+
+      expect(jsonResponse).to.have.property('version');
+      expect(jsonResponse.version).to.be.a('string');
+      expect(jsonResponse.version).to.be.equal(moduleInfo.version);
+    });
+
+    it('should check the JSON Response : `enabled`', async function () {
+      await testContext.addContextItem(this, 'testIdentifier', 'checkResponseEnabled', baseContext);
+
+      expect(jsonResponse).to.have.property('enabled');
+      expect(jsonResponse.enabled).to.be.a('boolean');
+      expect(jsonResponse.enabled).to.be.equal(moduleInfo.enabled);
     });
   });
