+++ tests/UI/campaigns/functional/API/02_endpoints/10_product/04_patchProductId.ts
@@ -24,7 +24,7 @@ import ProductData from '@data/faker/product';
 import {expect} from 'chai';
 import type {APIRequestContext, BrowserContext, Page} from 'playwright';
 
-const baseContext: string = 'functional_API_endpoints_product_patchAPIProductId';
+const baseContext: string = 'functional_API_endpoints_product_patchProductId';
 
 describe('API : PATCH /product/{productId}', async () => {
   let apiContext: APIRequestContext;
